Card #37. Added all of the logic to one module instead of using multiple
