--- modules/custom/book_show/book_show.info.yml
@@ -1,6 +1,0 @@
-name: book_show
-type: module
-description: The description.
-package: Custom
-core: 8.x
-core_version_requirement: ^8 || ^9
