--- modules/custom/book_show/book_show.module
@@ -1,19 +1,0 @@
-<?php
-
-/**
- * @file
- * Primary module hooks for book_show module.
- *
- * @DCG
- * This file is no longer required in Drupal 8.
- * @see https://www.drupal.org/node/2217931
- */
-
-function book_show_theme($existing,$type,$theme,$path){
-    
-    return [
-        'data_list' => array(
-            'variables' => array('items' => array(), 'title' => '')
-        )
-    ];
-}
