--- modules/custom/data_import/data_import.info.yml
@@ -1,6 +1,0 @@
-name: data_import
-type: module
-description: The description.
-package: Custom
-core: 8.x
-core_version_requirement: ^8 || ^9
