--- modules/custom/data_import/data_import.module
@@ -1,12 +1,0 @@
-<?php
-
-/**
- * @file
- * Primary module hooks for data_import module.
- *
- * @DCG
- * This file is no longer required in Drupal 8.
- * @see https://www.drupal.org/node/2217931
- */
-
-
