--- modules/custom/movie_export_controller/movie_export_controller.module
@@ -1,10 +1,0 @@
-<?php
-
-function movie_export_controller_theme($existing,$type,$theme,$path){
-    
-    return [
-        'export_list' => array(
-            'variables' => array('items' => array(), 'title' => '')
-        )
-    ];
-}
