--- modules/custom/movie_export_controller/src/Controller/MovieExportController.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace Drupal\movie_export_controller\Controller;
-
-    class MovieExportController {
-
-
-        public function page(){
-
-            $items = [
-                ['name' => 'Article one'],
-                ['name' => 'Article two'],
-                ['name' => 'Article three'],
-            ];
-
-            return [
-               '#theme' => 'export_list',
-               '#items' => $items,
-               '#title' => 'Our movie export controller title' 
-            ];
-        }
-
-    }
