+++ modules/custom/movie_reservation_controller/movie_reservation_controller.module
@@ -5,6 +5,12 @@ function movie_reservation_controller_theme($existing,$type,$theme,$path){
     return [
         'article_list' => array(
             'variables' => array('items' => array(), 'title' => '')
+        ),
+        'data_list' => array(
+            'variables' => array('items' => array(), 'title' => '')
+        ),
+        'export_list' => array(
+            'variables' => array('items' => array(), 'title' => '')
         )
     ];
 }
