+++ modules/custom/movie_reservation_controller/templates/article-list.html.twig
@@ -1,14 +1,13 @@
 <h4>{{ title}}</h4>
 
-{{ attach_library('movie_type_filter_controller/form-styling') }}
+{{ attach_library('movie_reservation_controller/form-styling') }}
 
 
 <div class="popup" id="popup">
     <h1 class="popup_title">Select date for movie</h1>
 
     <div id="popup_response_message"></div>
-
-    <input type="text" id="customer_name" name="customer_name" placeholder="Please enter your name"><br>
+     <input type="text" id="customer_name" name="customer_name" placeholder="Please enter your name"><br>
     
 
     <button type="button" class="cancel_btn" id="cancel">Cancel</button>
