--- modules/custom/movie_type_filter_controller/movie_type_filter_controller.info.yml
@@ -1,4 +1,0 @@
-name: movie type filter
-description: Movie type filter controller which will select all movies of a certail genre
-type: module
-core_version_requirement: ^8 || ^9
