--- modules/custom/movie_type_filter_controller/movie_type_filter_controller.module
@@ -1,10 +1,0 @@
-<?php
-
-function movie_type_filter_controller_theme($existing,$type,$theme,$path){
-    
-    return array(
-        'movie-list' => array(
-            'variables' => array('items' => array(), 'title' => '')
-        )
-    );
-}
