--- modules/custom/reservation_upload/reservation_upload.info.yml
@@ -1,6 +1,0 @@
-name: reservation_upload
-type: module
-description: The module which will be used to upload reservations to the Reservation database
-package: Custom
-core: 8.x
-core_version_requirement: ^8 || ^9
