--- modules/custom/reservation_upload/reservation_upload.module
@@ -1,10 +1,0 @@
-<?php
-
-/**
- * @file
- * Primary module hooks for reservation_upload module.
- *
- * @DCG
- * This file is no longer required in Drupal 8.
- * @see https://www.drupal.org/node/2217931
- */
