--- modules/custom/reservation_upload/reservation_upload.routing.yml
@@ -1,8 +1,0 @@
-reservation_upload.example:
-  path: '/reservation-upload/{customer_name}/{movie}'
-  defaults:
-    _title: 'Example'
-    _controller: '\Drupal\reservation_upload\Controller\ReservationUploadController::build'
-  methods:  [GET]
-  requirements:
-    _access: 'TRUE'
